Refactor: Persist and manage documents in Isar DB
This commit introduces document persistence using IsarDB. Documents (images and their extracted total expense) are now saved to the `MainPocket` within the `User` schema. The `DocumentBloc` has been updated to interact with an `IUserStorageRepository` for these operations.

Key changes:
- **Isar Integration for Documents:**
    - `UserStorageRepository` now implements `IUserStorageRepository`.
    - Added `addDocumentToMainPocket` to save a document's image (as `Uint8List`), filename (ID), and total expense to the user's `mainPocket.expenses`.
    - Added `deleteDocumentFromMainPocketByFileName` to remove a document from `mainPocket.expenses` based on its filename.
    - `Document` schema's `image` field type changed from `List<int>` to `List<byte>` for storing image data.
- **DocumentBloc Enhancements:**
    - `DocumentBloc` now takes `IUserStorageRepository` as a dependency.
    - After successfully processing an image with Gemini and extracting the total expense, the document (image and metadata) is saved using `_userStorageRepository.addDocumentToMainPocket`.
    - When a document is deleted via the `removeImage` event, it's also removed from Isar using `_userStorageRepository.deleteDocumentFromMainPocketByFileName`.
- **Image Processing Refinements:**
    - Removed image compression logic (`compressImage` utility and related stopwatch/logging) from `DocumentBloc`. Images are now sent to Gemini in their original format.
    - Simplified image data handling, directly reading `itemToProcess.file!.readAsBytes()` for Gemini.
- **File System Management:**
    - Added `removeImageFromMainPocketDirectory` utility in `Document/utils.dart` to delete image files from the local application documents directory when they are removed from the Isar DB. (Note: This is not yet explicitly called in this commit).
- **Service Locator Updates:**
    - `setupServiceLocator` is now an `async` function.
    - `DocumentBloc` is now registered with its `UserStorageRepository` dependency in `service_locator.dart`.
- **Core Exports:**
    - Updated `lib/core/index.dart` to export `UserStorageRepository`, `IUserStorageRepository`, and `service_locator.dart`.
- **Main.dart:**
    - Added try-catch around `setupServiceLocator()` call for better error handling during initialization.

+++ lib/core/Schema/User.dart
@@ -52,5 +52,5 @@ class Document {
   String totalExpense = '0';
 
   @Name("image")
-  List<int> image = [];
+  List<byte> image = [];
 }
